requirements.txt
APScheduler==3.11.0
asgiref==3.8.1
async-timeout==5.0.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2.5
django-apscheduler==0.7.0
django-cors-headers==4.2.0
django-redis==5.4.0
djangorestframework==3.14.0
idna==3.10
psycopg2-binary==2.9.7
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
requests==2.32.2
sqlparse==0.5.2
termcolor==2.4.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3